Combined all DTO test classes into one, fixed bug in EndPoint where url would append after one url build. Fixed WindDTO runtime error with gust.
